Split up API and Mock Api - Will need to refactor code to use reducers - See src/app/api/index for the todo
