+++ src/app/infrastructure/arrayToMap.ts
@@ -1,0 +1,7 @@
+export default function arrayToMap<T, TKey>(array: T[], keySelector: (t: T) => TKey) {
+    const mappedArray = array.reduce<any>((map, i) => {
+        map[keySelector(i)] = i;
+        return map;
+    }, {});
+    return mappedArray;
+}
